expliciting some twitch emojis
Added common use Twitch emojis, so the bot can use them more often. The previous version did not work so well, making it use the explicit VirtualHug in all the sentences. Incrementing the same command to the Twitch global emotes should fix that.

Expliciting the exact feeling in each of the emojis can help the bot to use i in the right contet instead of random context